requirements.txt
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
fastapi==0.120.2
filelock==3.20.0
fsspec==2025.9.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
llvmlite==0.45.1
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
packaging==25.0
pluggy==1.6.0
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
pytest-mock==3.15.1
python-multipart==0.0.20
regex==2025.10.23
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
starlette==0.49.1
sympy==1.14.0
tiktoken==0.12.0
torch==2.9.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
